+++ src/systems/cable-manual-wt.ts
@@ -1,8 +1,4 @@
-// Outputs sheet: Cable Manual WT/Log column
-import { Chipping } from '../methods/chipping';
-import { Loading } from '../methods/loading';
-import { calculateMachineCosts } from '../methods/machinecosts';
-import { MoveInCosts } from '../methods/moveincost';
+// Outputs sheet: Cable Manual WT column
 import {
   Assumptions,
   FrcsInputs,
@@ -10,11 +6,16 @@ import {
   IntermediateVariables,
   MachineCosts,
 } from '../model';
+import { Chipping } from './methods/chipping';
 import { CYCCU } from './methods/cyccu';
 import { CYPCU } from './methods/cypcu';
-import { FellwtChipLogOther } from './methods/fellwtchiplogother';
+import { FellwtSmallLogOther } from './methods/fellwtsmalllogother';
+import { Loading } from './methods/loading';
+import { calculateMachineCosts } from './methods/machinecosts';
+import { MoveInCosts } from './methods/moveincost';
+import { Processing } from './methods/processing';
 
-export function cableManualWTLog(
+export function cableManualWT(
   input: FrcsInputs,
   intermediate: IntermediateVariables,
   assumption: Assumptions
@@ -24,27 +25,23 @@ export function cableManualWTLog(
   const BoleVolCCF = intermediate.volPerAcre / 100;
   const ResidueRecoveredPrimary = assumption.ResidueRecovFracWT * intermediate.residueCT;
   const PrimaryProduct = intermediate.boleWeight + ResidueRecoveredPrimary;
-  const ResidueRecoveredOptional = 0;
+  const ResidueRecoveredOptional = input.includeCostsCollectChipResidues
+    ? assumption.ResidueRecovFracWT * intermediate.residueSLT
+    : 0;
   const TotalPrimaryProductsAndOptionalResidues = PrimaryProduct + ResidueRecoveredOptional;
-  // Amounts Unrecovered and Left within the Stand Per Acre
-  const GroundFuel =
-    intermediate.residueALT + intermediate.residueCT * (1 - assumption.ResidueRecovFracWT);
-  // Amounts Unrecovered and Left at the Landing
-  const PiledFuel = 0;
-  // TotalResidues
-  const ResidueUncutTrees = 0;
-  const TotalResidues =
-    ResidueRecoveredPrimary + ResidueRecoveredOptional + ResidueUncutTrees + GroundFuel + PiledFuel;
+
   // Machine costs
   const machineCost: MachineCosts = calculateMachineCosts(input.dieselFuelPrice);
   // System Cost Elements-------
-  const FellwtChipLogOtherResults = FellwtChipLogOther(input, intermediate, machineCost);
-  const CostManFLBALT2 = FellwtChipLogOtherResults.CostManFLBALT2;
-  const CostManFellCT2 = FellwtChipLogOtherResults.CostManFellCT2;
+  const FellwtSmallLogOtherResults = FellwtSmallLogOther(input, intermediate, machineCost);
+  const CostManFLBLLT2 = FellwtSmallLogOtherResults.CostManFLBLLT2;
+  const CostManFellST2 = FellwtSmallLogOtherResults.CostManFellST2;
   const CYPCUresults = CYPCU(assumption, input, intermediate, machineCost);
   const CostYardPCUB = CYPCUresults.CostYardPCUB;
   const CYCCUresults = CYCCU(input, intermediate, machineCost);
   const CostYardCCUB = CYCCUresults.CostYardCCUB;
+  const ProcessingResults = Processing(input, intermediate, machineCost);
+  const CostProcess = ProcessingResults.CostProcess;
   const LoadingResults = Loading(assumption, input, intermediate, machineCost);
   const CostLoad = LoadingResults.CostLoad;
   const ChippingResults = Chipping(assumption, input, intermediate, machineCost);
@@ -60,12 +57,14 @@ export function cableManualWTLog(
   const GalChainsaw = 0.0104 * 2.83168 * 0.264172; // 0.0104 L/m3 => gal/CCF
   const GalYardPCUB = CYPCUresults.GalYardPCUB;
   const GalYardCCUB = CYCCUresults.GalYardCCUB;
+  const GalProcess = ProcessingResults.GalProcess;
   const GalLoad = LoadingResults.GalLoad;
   const GalChipWT = ChippingResults.GalChipWT;
+  const GalChipLooseRes = ChippingResults.GalChipLooseRes;
 
   // C. For All Products, $/ac
-  const ManualFellLimbBuckAllLogTrees = (CostManFLBALT2 * intermediate.volPerAcreALT) / 100;
-  const ManualFellChipTrees = (CostManFellCT2 * intermediate.volPerAcreCT) / 100;
+  const ManualFellLimbBuckTreesLarger80cf = (CostManFLBLLT2 * intermediate.volPerAcreLLT) / 100;
+  const ManualFellTreesLess80cf = (CostManFellST2 * intermediate.volPerAcreST) / 100;
   const CableYardUnbunchedAllTrees =
     ((input.isPartialCut === true
       ? CostYardPCUB
@@ -74,6 +73,7 @@ export function cableManualWTLog(
       : 0) *
       intermediate.volPerAcre) /
     100;
+  const ProcessLogTreesLess80cf = (CostProcess * intermediate.volPerAcreSLT) / 100;
   const LoadLogTrees = (CostLoad * intermediate.volPerAcreALT) / 100;
   const ChipWholeTrees =
     (CostChipWT *
@@ -81,21 +81,22 @@ export function cableManualWTLog(
     100;
 
   const Stump2Truck4PrimaryProductWithoutMovein =
-    ManualFellLimbBuckAllLogTrees +
-    ManualFellChipTrees +
+    ManualFellLimbBuckTreesLarger80cf +
+    ManualFellTreesLess80cf +
     CableYardUnbunchedAllTrees +
-    (input.isBiomassSalvage === false ? LoadLogTrees : 0) +
+    (input.isBiomassSalvage === false ? ProcessLogTreesLess80cf + LoadLogTrees : 0) +
     ChipWholeTrees;
   const Stump2Truck4ResiduesWithoutMovein =
     ChipWholeTrees +
-    ManualFellChipTrees +
+    ManualFellTreesLess80cf *
+      ((intermediate.boleWeightCT + intermediate.residueCT) /
+        (intermediate.boleWeightST + intermediate.residueST)) +
     CableYardUnbunchedAllTrees *
       ((intermediate.boleWeightCT + intermediate.residueCT) /
         (intermediate.boleWeight + intermediate.residue));
   const Movein4PrimaryProduct = input.includeMoveInCosts
-    ? MoveInCostsResults.CostPerCCFcableManualWTLog * BoleVolCCF
+    ? MoveInCostsResults.CostPerCCFcableManualWT * BoleVolCCF
     : 0;
-
   const ChipLooseResiduesFromLogTreesLess80cf = input.includeCostsCollectChipResidues
     ? CostChipLooseRes * ResidueRecoveredOptional
     : 0;
@@ -106,12 +107,13 @@ export function cableManualWTLog(
       : 0;
 
   // D. For All Products, gal/ac
-  const ManualFellLimbBuckAllLogTrees2 = (GalChainsaw * intermediate.volPerAcreALT) / 100;
-  const ManualFellChipTrees2 = (GalChainsaw * intermediate.volPerAcreCT) / 100;
+  const ManualFellLimbBuckTreesLarger80cf2 = (GalChainsaw * intermediate.volPerAcreLLT) / 100;
+  const ManualFellTreesLess80cf2 = (GalChainsaw * intermediate.volPerAcreST) / 100;
   const CableYardUnbunchedAllTrees2 =
     ((input.isPartialCut === true ? GalYardPCUB : input.isPartialCut === false ? GalYardCCUB : 0) *
       intermediate.volPerAcre) /
     100;
+  const ProcessLogTreesLess80cf2 = (GalProcess * intermediate.volPerAcreSLT) / 100;
   const LoadLogTrees2 = (GalLoad * intermediate.volPerAcreALT) / 100;
   const ChipWholeTrees2 =
     (GalChipWT *
@@ -120,21 +122,28 @@ export function cableManualWTLog(
 
   const DieselStump2Truck4PrimaryProductWithoutMovein =
     CableYardUnbunchedAllTrees2 +
-    (input.isBiomassSalvage === false ? LoadLogTrees2 : 0) +
-    +ChipWholeTrees2;
+    (input.isBiomassSalvage === false ? ProcessLogTreesLess80cf2 + LoadLogTrees2 : 0) +
+    ChipWholeTrees2;
   const DieselStump2Truck4ResiduesWithoutMovein =
     CableYardUnbunchedAllTrees2 *
       ((intermediate.boleWeightCT + intermediate.residueCT) /
         (intermediate.boleWeight + intermediate.residue)) +
     ChipWholeTrees2;
   const GasolineStump2Truck4PrimaryProductWithoutMovein =
-    ManualFellLimbBuckAllLogTrees2 + ManualFellChipTrees2;
-  const GasolineStump2Truck4ResiduesWithoutMovein = ManualFellChipTrees2;
-  const LowboyLoads = 4;
+    ManualFellLimbBuckTreesLarger80cf2 + ManualFellTreesLess80cf2;
+  const GasolineStump2Truck4ResiduesWithoutMovein =
+    ManualFellTreesLess80cf2 *
+    ((intermediate.boleWeightCT + intermediate.residueCT) /
+      (intermediate.boleWeightST + intermediate.residueST));
+  const LowboyLoads = 5;
   const mpg = 6;
   const Movein4PrimaryProduct2 = input.includeMoveInCosts
     ? (LowboyLoads * input.moveInDistance) / mpg / input.area
     : 0;
+  const ChipLooseResiduesFromLogTreesLess80cf2 = input.includeCostsCollectChipResidues
+    ? GalChipLooseRes * ResidueRecoveredOptional
+    : 0;
+  const OntoTruck4ResiduesWoMovein2 = ChipLooseResiduesFromLogTreesLess80cf2;
 
   // III. Summaries
   const frcsOutputs: FrcsOutputs = {
@@ -173,7 +182,9 @@ export function cableManualWTLog(
   frcsOutputs.total.costPerGT = frcsOutputs.total.costPerAcre / frcsOutputs.total.yieldPerAcre;
   // Fuel
   frcsOutputs.total.dieselPerAcre =
-    DieselStump2Truck4PrimaryProductWithoutMovein + Movein4PrimaryProduct2;
+    DieselStump2Truck4PrimaryProductWithoutMovein +
+    Movein4PrimaryProduct2 +
+    OntoTruck4ResiduesWoMovein2;
   frcsOutputs.total.dieselPerBoleCCF = frcsOutputs.total.dieselPerAcre / BoleVolCCF;
   frcsOutputs.total.gasolinePerAcre = GasolineStump2Truck4PrimaryProductWithoutMovein;
   frcsOutputs.total.gasolinePerBoleCCF = frcsOutputs.total.gasolinePerAcre / BoleVolCCF;
@@ -186,7 +197,8 @@ export function cableManualWTLog(
   frcsOutputs.biomass.costPerBoleCCF = frcsOutputs.biomass.costPerAcre / BoleVolCCF;
   frcsOutputs.biomass.costPerGT = frcsOutputs.biomass.costPerAcre / frcsOutputs.total.yieldPerAcre;
   // Fuel
-  frcsOutputs.biomass.dieselPerAcre = DieselStump2Truck4ResiduesWithoutMovein;
+  frcsOutputs.biomass.dieselPerAcre =
+    DieselStump2Truck4ResiduesWithoutMovein + OntoTruck4ResiduesWoMovein2;
   frcsOutputs.biomass.gasolinePerAcre = GasolineStump2Truck4ResiduesWithoutMovein;
 
   if (input.isBiomassSalvage) {
