+++ src/systems/helicopter-ctl.ts
@@ -1,9 +1,4 @@
-// Outputs sheet: Cable CTL column
-import { Chipping } from '../methods/chipping';
-import { Harvesting } from '../methods/harvesting';
-import { Loading } from '../methods/loading';
-import { calculateMachineCosts } from '../methods/machinecosts';
-import { MoveInCosts } from '../methods/moveincost';
+// Outputs sheet: Helicopter CTL column
 import {
   Assumptions,
   FrcsInputs,
@@ -11,9 +6,14 @@ import {
   IntermediateVariables,
   MachineCosts,
 } from '../model';
-import { CYCTL } from './methods/cyctl';
+import { Chipping } from './methods/chipping';
+import { Harvesting } from './methods/harvesting';
+import { HelicopterYarding } from './methods/helicopteryarding';
+import { Loading } from './methods/loading';
+import { calculateMachineCosts } from './methods/machinecosts';
+import { MoveInCosts } from './methods/moveincost';
 
-export function cableCTL(
+export function helicopterCTL(
   input: FrcsInputs,
   intermediate: IntermediateVariables,
   assumption: Assumptions
@@ -31,10 +31,9 @@ export function cableCTL(
   // System Cost Elements-------
   const HarvestingResults = Harvesting(assumption, input, intermediate, machineCost);
   const CostHarvest = HarvestingResults.CostHarvest;
-  const CYCTLresults = CYCTL(assumption, input, intermediate, machineCost);
-  const CostYardCTL = CYCTLresults.CostYardCTL;
-  const LoadingResults = Loading(assumption, input, intermediate, machineCost);
-  const CostLoadCTL = LoadingResults.CostLoadCTL;
+  const HelicopterYardingResults = HelicopterYarding(input, intermediate);
+  const CostHeliYardCTL = HelicopterYardingResults.CostHeliYardCTL;
+  const CostHeliLoadCTL = HelicopterYardingResults.CostHeliLoadCTL;
   const ChippingResults = Chipping(assumption, input, intermediate, machineCost);
   const CostChipWT = ChippingResults.CostChipWT;
   const MoveInCostsResults = MoveInCosts(
@@ -46,14 +45,16 @@ export function cableCTL(
   const CostChipBundledRes = ChippingResults.CostChipBundledRes;
 
   const GalHarvest = HarvestingResults.GalHarvest;
-  const GalYardCTL = CYCTLresults.GalYardCTL;
+  const GalHeliYardCTL = HelicopterYardingResults.GalHeliYardCTL;
+  const LoadingResults = Loading(assumption, input, intermediate, machineCost);
   const GalLoadCTL = LoadingResults.GalLoadCTL;
+  const GalHeliLoadCTL = GalLoadCTL * 2;
   const GalChipWT = ChippingResults.GalChipWT;
 
   // C. For All Products, $/ac
   const HarvestTreesLess80cf = (CostHarvest * intermediate.volPerAcreST) / 100;
-  const CableYardCTLtreesLess80cf = (CostYardCTL * intermediate.volPerAcreST) / 100;
-  const LoadCTLlogTreesLess80cf = (CostLoadCTL * intermediate.volPerAcreSLT) / 100;
+  const HeliYardCTLtreesLess80cf = (CostHeliYardCTL * intermediate.volPerAcreST) / 100;
+  const LoadCTLlogTreesLess80cf = (CostHeliLoadCTL * intermediate.volPerAcreSLT) / 100;
   const ChipTreeBoles =
     (CostChipWT *
       (input.isBiomassSalvage === false ? intermediate.volPerAcreCT : intermediate.volPerAcreST)) /
@@ -61,15 +62,15 @@ export function cableCTL(
 
   const Stump2Truck4PrimaryProductWithoutMovein =
     HarvestTreesLess80cf +
-    CableYardCTLtreesLess80cf +
+    HeliYardCTLtreesLess80cf +
     (input.isBiomassSalvage === false ? LoadCTLlogTreesLess80cf : 0) +
     ChipTreeBoles;
   const Stump2Truck4ResiduesWithoutMovein =
     ChipTreeBoles +
-    (HarvestTreesLess80cf + CableYardCTLtreesLess80cf) *
+    (HarvestTreesLess80cf + HeliYardCTLtreesLess80cf) *
       (intermediate.boleWeightCT / intermediate.boleWeightST);
   const Movein4PrimaryProduct = input.includeMoveInCosts
-    ? MoveInCostsResults.CostPerCCFcableCTL * BoleVolCCF
+    ? MoveInCostsResults.CostPerCCFheliCTL * BoleVolCCF
     : 0;
   const ChipBundledResiduesFromTreesLess80cf = input.includeCostsCollectChipResidues
     ? CostChipBundledRes * ResidueRecoveredOptional
@@ -77,13 +78,13 @@ export function cableCTL(
   const OntoTruck4ResiduesWoMovein = ChipBundledResiduesFromTreesLess80cf;
   const Movein4Residues =
     input.includeMoveInCosts && input.includeCostsCollectChipResidues
-      ? MoveInCostsResults.CostPerCCFbundleResidues * ResidueRecoveredOptional
+      ? MoveInCostsResults.CostPerCCFheliCTL * ResidueRecoveredOptional
       : 0;
 
-  // D. For All Products, gal/ac
+  // D. For All Products, $/ac
   const HarvestTreesLess80cf2 = (GalHarvest * intermediate.volPerAcreST) / 100;
-  const CableYardCTLtreesLess80cf2 = (GalYardCTL * intermediate.volPerAcreST) / 100;
-  const LoadCTLlogTreesLess80cf2 = (GalLoadCTL * intermediate.volPerAcreSLT) / 100;
+  const HeliYardCTLtreesLess80cf2 = (GalHeliYardCTL * intermediate.volPerAcreST) / 100;
+  const LoadCTLlogTreesLess80cf2 = (GalHeliLoadCTL * intermediate.volPerAcreSLT) / 100;
   const ChipTreeBoles2 =
     (GalChipWT *
       (input.isBiomassSalvage === false ? intermediate.volPerAcreCT : intermediate.volPerAcreST)) /
@@ -91,18 +92,17 @@ export function cableCTL(
 
   const DieselStump2Truck4PrimaryProductWithoutMovein =
     HarvestTreesLess80cf2 +
-    CableYardCTLtreesLess80cf2 +
     (input.isBiomassSalvage === false ? LoadCTLlogTreesLess80cf2 : 0) +
     ChipTreeBoles2;
-  const DieselStump2Truck4ResiduesWithoutMovein =
-    (HarvestTreesLess80cf2 + CableYardCTLtreesLess80cf2) *
-      (intermediate.boleWeightCT / intermediate.boleWeightST) +
-    ChipTreeBoles2;
+  const DieselStump2Truck4ResiduesWithoutMovein = ChipTreeBoles2;
   const LowboyLoads = 4;
   const mpg = 6;
   const Movein4PrimaryProduct2 = input.includeMoveInCosts
     ? (LowboyLoads * input.moveInDistance) / mpg / input.area
     : 0;
+  const JetFuelStump2Truck4PrimaryProductWithoutMovein = HeliYardCTLtreesLess80cf2;
+  const JetFuelStump2Truck4ResiduesWithoutMovein =
+    HeliYardCTLtreesLess80cf2 * (intermediate.boleWeightCT / intermediate.boleWeightST);
 
   // III. Summaries
   const frcsOutputs: FrcsOutputs = {
@@ -143,6 +143,8 @@ export function cableCTL(
   frcsOutputs.total.dieselPerAcre =
     DieselStump2Truck4PrimaryProductWithoutMovein + Movein4PrimaryProduct2;
   frcsOutputs.total.dieselPerBoleCCF = frcsOutputs.total.dieselPerAcre / BoleVolCCF;
+  frcsOutputs.total.jetFuelPerAcre = JetFuelStump2Truck4PrimaryProductWithoutMovein;
+  frcsOutputs.total.jetFuelPerBoleCCF = frcsOutputs.total.jetFuelPerAcre / BoleVolCCF;
 
   // System Summaries - Residue
   // Cost
@@ -154,6 +156,7 @@ export function cableCTL(
   frcsOutputs.biomass.costPerGT = frcsOutputs.biomass.costPerAcre / frcsOutputs.total.yieldPerAcre;
   // Fuel
   frcsOutputs.biomass.dieselPerAcre = DieselStump2Truck4ResiduesWithoutMovein;
+  frcsOutputs.biomass.jetFuelPerAcre = JetFuelStump2Truck4ResiduesWithoutMovein;
 
   if (input.isBiomassSalvage) {
     frcsOutputs.biomass = frcsOutputs.total;
