+++ src/systems/methods/chipping.ts
@@ -1,5 +1,5 @@
 // Chipping sheet
-import { Assumptions, FrcsInputs, IntermediateVariables, MachineCosts } from '../model';
+import { Assumptions, FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function Chipping(
   assumption: Assumptions,
