+++ src/systems/methods/fellalltrees.ts
@@ -1,5 +1,5 @@
 // Felling (all trees) sheet
-import { FrcsInputs, IntermediateVariables, MachineCosts } from '../model';
+import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function FellAllTrees(
   input: FrcsInputs,
