+++ src/systems/methods/felllargelogtrees.ts
@@ -1,5 +1,5 @@
 // FellLargeLogTrees sheet
-import { FrcsInputs, IntermediateVariables, MachineCosts } from '../model';
+import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function FellLargeLogTrees(
   input: FrcsInputs,
