+++ src/systems/methods/processing.ts
@@ -1,5 +1,5 @@
 // Processing sheet
-import { FrcsInputs, IntermediateVariables, MachineCosts } from '../model';
+import { FrcsInputs, IntermediateVariables, MachineCosts } from '../../model';
 
 function Processing(
   input: FrcsInputs,
